# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/assistant/message": {
      "post": {
        "operationId": "sendAssistantMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantReplyDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Assistant"
        ]
      }
    },
    "/api/assistant/confirm": {
      "post": {
        "operationId": "confirmAssistantDraft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantConfirmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantReplyDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Assistant"
        ]
      }
    },
    "/api/assistant/cancel": {
      "post": {
        "operationId": "cancelAssistantDraft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantConfirmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantReplyDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Assistant"
        ]
      }
    },
...
  },
  "info": {
    "title": "Subsio API",
    "description": "API for subscriptions, recurring payments and reminders",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sm_session"
      }
    },
    "schemas": {
...
      "AssistantMessageDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1
          }
        },
        "required": [
          "message"
        ]
      },
      "AssistantDraftDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "billingPeriod": {
            "type": "string",
            "nullable": true
          },
          "nextChargeDate": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodLabel": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id"
        ]
      },
      "AssistantReplyDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "message",
              "draft",
              "created"
            ]
          },
          "message": {
            "type": "string"
          },
          "draft": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AssistantDraftDto"
              }
            ]
          }
        },
        "required": [
          "kind",
          "message"
        ]
      },
      "AssistantConfirmDto": {
        "type": "object",
        "properties": {
          "draftId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "draftId"
        ]
      },
...
    }
  }
}