# packages/api-client/src/generated/models/assistantConfirmDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Subsio API
 * API for subscriptions, recurring payments and reminders
 * OpenAPI spec version: 0.1.0
 */

export interface AssistantConfirmDto {
  draftId: string;
}