# packages/api-client/src/generated/models/assistantDraftDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Subsio API
 * API for subscriptions, recurring payments and reminders
 * OpenAPI spec version: 0.1.0
 */

export interface AssistantDraftDto {
  id: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  billingPeriod?: string | null;
  /** @nullable */
  nextChargeDate?: string | null;
  /** @nullable */
  paymentMethodLabel?: string | null;
}