# packages/api-client/src/generated/models/assistantMessageDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Subsio API
 * API for subscriptions, recurring payments and reminders
 * OpenAPI spec version: 0.1.0
 */

export interface AssistantMessageDto {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  message: string;
}