# packages/api-client/src/generated/models/assistantReplyDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Subsio API
 * API for subscriptions, recurring payments and reminders
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantDraftDto } from './assistantDraftDto';
import type { AssistantReplyDtoKind } from './assistantReplyDtoKind';

export interface AssistantReplyDto {
  kind: AssistantReplyDtoKind;
  message: string;
  /** @nullable */
  draft?: AssistantDraftDto | null;
}